Error early if notarization username/password is not set
(cherry picked from commit 082f87358850f70e22f74a6290dbef58810a0f42)

+++ create_osx_installer.sh
@@ -13,6 +13,16 @@ fi
 
 
 if !($development); then
+	if [ "x$AC_USERNAME" = x ]; then
+		echo "Notarization username not set (AC_USERNAME)"
+		exit 1
+	fi
+
+	if [ "x$AC_PASSWORD" = x ]; then
+		echo "Notarization account password not set (AC_PASSWORD)"
+		exit 1
+	fi
+
 	git reset --hard
 	cd client
 	git reset --hard
